refactor: delegate input checks to private methods in userService

+++ src/main/java/app/repository/UserRepository.java
@@ -10,12 +10,12 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     @Query("SELECT u FROM User u WHERE u.dni = ?1")
     Optional<User> findUserByDni(String dni);
 
     @Transactional
     @Modifying
     @Query("UPDATE User u set u.dni = ?1, u.firstName = ?2, u.lastName = ?3, u.password = ?4 WHERE u.userId = ?5")
-    void updateUserById(String dni, String firstName, String lastName, String password, Integer userId);
+    void updateUserById(String dni, String firstName, String lastName, String password, Long userId);
 }
